+++ jest.config.js
@@ -12,6 +12,7 @@ export default {
   preset: "ts-jest/presets/default-esm",
   modulePathIgnorePatterns: ["src/__tests__/generated", "src/generated", "dist"],
   coverageDirectory: "reports/jest-coverage",
+  collectCoverage: true,
   moduleFileExtensions: ["js", "json", "ts", "d.ts"],
   reporters: ["default"],
   modulePaths: [compilerOptions.baseUrl],